Feature: Added Context for Wallet Connection criteria
