+++ components/contexts/WalletConnected.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const WalletConnected = createContext();
+
+export default WalletConnected;
