+++ contracts/vesting.js
@@ -1,4 +1,5 @@
-const contractAddress = "0x828e4ED5a77aDED293E77C9D3e9CFb815B49a006";
+// const contractAddress = "0x828e4ED5a77aDED293E77C9D3e9CFb815B49a006";
+const contractAddress = "0x5f17b59FCDb08Bc562368031E4414F66769e6152";
 const abi = [
   {
     inputs: [],
