+++ pages/Mint.jsx
@@ -1,6 +1,5 @@
-import Header from '@/components/Header'
 import React from 'react'
-import { useState, useEffect } from "react";
+import { useState, useEffect, useContext } from "react";
 import { ethers } from 'ethers';
 import {
     useAccount,
@@ -14,9 +13,12 @@ import { abi, contractAddress } from "../contracts/vesting";
 import HoldersList from '@/components/HoldersList';
 import MintList from '@/components/MintList';
 
+import WalletConnected from '@/components/contexts/WalletConnected';
+import ConnectTheWallet from '@/components/ConnectTheWallet';
+
 function Mint() {
     const [shouldReinitialize, setShouldReinitialize] = useState(false);
-
+    const accountConnection = useContext(WalletConnected);
 
     const [role, setRole] = useState();
     const [stakeHolderAddress, setStakeHolderAddress] = useState();
@@ -111,49 +113,54 @@ function Mint() {
 
     return (
         <div>
-            <Header />
-
-            <div className='mx-20 my-10 bg-slate-300  px-4 py-2 rounded-lg shadow-md hover:shadow-lg overflow-hidden'>
-                <h1 className='text-3xl font-bold mt-10 mb-5'>Mint Tokens</h1>
-                <form className='flex flex-col space-y-4 mb-10'>
-                    <label className='flex flex-col'>
-                        <span className='mb-1 font-bold'>Role:</span>
-                        <input type="text" value={role} onChange={handleRoleChange} className='border border-gray-400 p-2 rounded-md' />
-                    </label>
-                    <label className='flex flex-col'>
-                        <span className='mb-1 font-bold'>Stake Holder Address:</span>
-                        <input type="text" value={stakeHolderAddress} onChange={handleStakeHolderAddress} className='border border-gray-400 p-2 rounded-md' />
-                    </label>
-                    <label className='flex flex-col'>
-                        <span className='mb-1 font-bold'>Tokens:</span>
-                        <input type="text" value={tokens} onChange={handleTokens} className='border border-gray-400 p-2 rounded-md' />
-                    </label>
-                    <label className='flex flex-col'>
-                        <span className='mb-1 font-bold'>Organisation Address:</span>
-                        <input type="text" value={organisationAddress} onChange={handleOrganisationAddress} className='border border-gray-400 p-2 rounded-md' />
-                    </label>
-                    <button type="submit" onClick={mintTokens} disabled={waitForTransaction.isLoading || registering} className='bg-blue-500 hover:bg-blue-600 text-white py-2 px-4 rounded-md'>{waitForTransaction.isLoading ? "Transacting..... " : (registering ? "Check Wallet" : "Register")}</button>
-                </form>
-            </div>
-
-            <div className='mx-20 my-10  bg-slate-300  px-4 py-2 rounded-lg shadow-md hover:shadow-lg overflow-hidden '>
-                <h1 className='text-3xl font-bold mt-10 mb-5'>Get Minters</h1>
-                <form className='flex flex-col space-y-4 mb-10'>
-                    <div className='flex flex-col'>
-                        <label htmlFor='role' className='mb-1 font-bold'>Role</label>
-                        <input type='text' id='role' value={roleRead} onChange={handleRoleReadChange} className='border border-gray-400 p-2 rounded-md' />
+            {accountConnection
+                ?
+                <>
+                    <div className='mx-20 my-10 bg-slate-300  px-4 py-2 rounded-lg shadow-md hover:shadow-lg overflow-hidden'>
+                        <h1 className='text-3xl font-bold mt-10 mb-5'>Mint Tokens</h1>
+                        <form className='flex flex-col space-y-4 mb-10'>
+                            <label className='flex flex-col'>
+                                <span className='mb-1 font-bold'>Role:</span>
+                                <input type="text" value={role} onChange={handleRoleChange} className='border border-gray-400 p-2 rounded-md' />
+                            </label>
+                            <label className='flex flex-col'>
+                                <span className='mb-1 font-bold'>Stake Holder Address:</span>
+                                <input type="text" value={stakeHolderAddress} onChange={handleStakeHolderAddress} className='border border-gray-400 p-2 rounded-md' />
+                            </label>
+                            <label className='flex flex-col'>
+                                <span className='mb-1 font-bold'>Tokens:</span>
+                                <input type="text" value={tokens} onChange={handleTokens} className='border border-gray-400 p-2 rounded-md' />
+                            </label>
+                            <label className='flex flex-col'>
+                                <span className='mb-1 font-bold'>Organisation Address:</span>
+                                <input type="text" value={organisationAddress} onChange={handleOrganisationAddress} className='border border-gray-400 p-2 rounded-md' />
+                            </label>
+                            <button type="submit" onClick={mintTokens} disabled={waitForTransaction.isLoading || registering} className='bg-blue-500 hover:bg-blue-600 text-white py-2 px-4 rounded-md'>{waitForTransaction.isLoading ? "Transacting..... " : (registering ? "Check Wallet" : "Register")}</button>
+                        </form>
                     </div>
-                    <div className='flex flex-col'>
-                        <label htmlFor='organisationAddress' className='mb-1 font-bold'>Organisation Address:</label>
-                        <input type='text' id='organisationAddress' value={organisationAddressRead} onChange={handleOrganisationAddressRead} className='border border-gray-400 p-2 rounded-md' />
-                    </div>
-                    <button type='submit' onClick={readHolders} className='bg-blue-500 hover:bg-blue-600 text-white py-2 px-4 rounded-md'>Get Minters</button>
-                </form>
-                {holders.toString() != "" &&
-                    <MintList holders={holders} org={organisationAddressRead} />
-                }
 
-            </div>
+                    <div className='mx-20 my-10  bg-slate-300  px-4 py-2 rounded-lg shadow-md hover:shadow-lg overflow-hidden '>
+                        <h1 className='text-3xl font-bold mt-10 mb-5'>Get Minters</h1>
+                        <form className='flex flex-col space-y-4 mb-10'>
+                            <div className='flex flex-col'>
+                                <label htmlFor='role' className='mb-1 font-bold'>Role</label>
+                                <input type='text' id='role' value={roleRead} onChange={handleRoleReadChange} className='border border-gray-400 p-2 rounded-md' />
+                            </div>
+                            <div className='flex flex-col'>
+                                <label htmlFor='organisationAddress' className='mb-1 font-bold'>Organisation Address:</label>
+                                <input type='text' id='organisationAddress' value={organisationAddressRead} onChange={handleOrganisationAddressRead} className='border border-gray-400 p-2 rounded-md' />
+                            </div>
+                            <button type='submit' onClick={readHolders} className='bg-blue-500 hover:bg-blue-600 text-white py-2 px-4 rounded-md'>Get Minters</button>
+                        </form>
+                        {holders.toString() != "" &&
+                            <MintList holders={holders} org={organisationAddressRead} />
+                        }
+
+                    </div>
+                </>
+                :
+                <ConnectTheWallet />
+            }
         </div>
     )
 }
