+++ pages/index.js
@@ -4,6 +4,7 @@ import styles from "@/styles/Home.module.css";
 import { ConnectButton } from "@rainbow-me/rainbowkit";
 import Header from "@/components/Header";
 import { useAccount } from "wagmi";
+import Register from "@/components/Register";
 
 export default function Home() {
   const { address } = useAccount();
@@ -16,12 +17,7 @@ export default function Home() {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <Header />
-      <main className={styles.main}>
-        <section className="center">
-          <Link href="/Register">{!address && <ConnectButton />}</Link>
-        </section>
-      </main>
+      <Register />
     </>
   );
 }
